test(agent): adding more tests
