+++ test/ai.spec.ts
@@ -1,8 +1,6 @@
 import { describe, expect } from "@jest/globals";
 import { embed, respond, summarize } from "@/lib/ai";
-import { createAgent, genCharacter } from "@/lib/database/agent";
-import { updateSettings } from "@/lib/database/settings";
-import { getSettings } from "@/lib/read";
+import { updateModelEmbeddingSetting } from "./test-helpers";
 
 describe("ai", () => {
   it("should be able to create an embedding using ollama", async () => {
@@ -33,32 +31,3 @@ describe("ai", () => {
     expect(output).toBeTruthy();
   });
 });
-
-describe("agent", () => {
-  it("should generate a character with default parameters", async () => {
-    const character = await genCharacter();
-    expect(character.name).toBeDefined();
-    expect(character.content).toBeDefined();
-    expect(character.combinedQualities).toBeDefined();
-  });
-
-  it("should be able to create an agent", async () => {
-    const name = "test_agent";
-    const result = await createAgent({ name });
-  });
-});
-
-describe("posts", () => {
-  it("should be able to create a post", async () => {
-  });
-});
-
-const updateModelEmbeddingSetting = async (newModelEmbedding: string) => {
-  const settings = await getSettings();
-  const modelEmbedding = settings.find((s) => s.name === "modelembedding");
-
-  if (modelEmbedding) {
-    modelEmbedding.defaultValue = newModelEmbedding;
-    await updateSettings(settings);
-  }
-};
