+++ test/log.mock.ts
@@ -1,0 +1,4 @@
+// import { Entity, Log } from "@/types/mod";
+
+export const createLog = jest.fn();
+export default createLog;
