+++ test/posts.spec.ts
@@ -1,0 +1,4 @@
+describe("posts", () => {
+  it("should be able to create a post", async () => {
+  });
+});
